packages/model-backend-py/pyproject.toml
[project]
name = "model-backend-py"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Daniel Imfeld", email = "daniel@imfeld.dev" }
]
dependencies = [
    "fastapi>=0.103.1",
    "transformers>=4.33.1",
    "python-multipart>=0.0.6",
    "uvicorn>=0.23.2",
    "torch>=2.0.1",
    "numpy>=1.23.0",
    "nemo_toolkit>=1.20.0",
    "hydra-core<1.3",
    "pytorch-lightning<=1.9.4",
    "librosa>=0.10.1",
    "pandas>=2.0.3",
    "sentencepiece<1.0.0",
    "youtokentome @ git+https://github.com/LahiLuk/YouTokenToMe",
    "inflect>=7.0.0",
    "webdataset<=0.1.62",
    "sacremoses>=0.0.43",
    "soundfile>=0.12.1",
    "sox>=1.4.1",
    "g2p_en>=2.1.0",
    "jiwer>=3.0.3",
    "pydub>=0.25.1",
    "scipy>=1.10.1",
    "pyannote.core>=5.0.0",
    "pyannote.metrics>=3.2.1",
    "editdistance>=0.6.2",
    "braceexpand>=0.1.7",
    "ipywidgets>=8.1.1",
    "kaldi-python-io>=1.2.2",
    "kaldiio>=2.18.0",
    "marshmallow>=3.20.1",
    "matplotlib>=3.7.3",
    "packaging>=23.1",
    "ruamel.yaml>=0.17.32"
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "black>=23.9.1",
]

[tool.hatch.metadata]
allow-direct-references = true
